SDA-4291 - Move interfaces into a dedicate file
