+++ src/app/auto-update-handler.ts
@@ -6,8 +6,8 @@ import { isMac, isWindowsOS } from '../common/env';
 import { logger } from '../common/logger';
 import { isUrl } from '../common/utils';
 import { whitelistHandler } from '../common/whitelist-handler';
-import { InstallActionTypes, InstallTypes } from './bi/analytics-handler';
 import { sendAutoUpdateAnalytics } from './bi/auto-update-analytics';
+import { InstallActionTypes, InstallTypes } from './bi/interface';
 import { config } from './config-handler';
 import { retrieveWindowsRegistry } from './registry-handler';
 import { EChannelRegistry, RegistryStore } from './stores/registry-store';
